// frontend/src/api/model/comment.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

/**
 * 评论实体
 */
export interface Comment {
  /** 评论ID */
  id?: number;
  /** 用户ID */
  userId?: string;
  /** 电影ID */
  movieId?: number;
  /** 评论标题 (长评专用) */
  title?: string;
  /** 评论类型 (1:短评, 2:长评) */
  type?: number;
  /** 评论内容 */
  content?: string;
  /** 获赞数 */
  votes?: number;
  /** 评论发布时间 */
  commentTime?: string;
  /** 乐观锁版本号 */
  version?: number;
}